Create a cli pal

A light wrapper around elmer chat functions to create a chat object with a custom system prompt.

Usage

cli_pal(
  fn = getOption(".clipal_fn", default = "new_chat_claude"),
  ...,
  .ns = "elmer"
)

Arguments

fn

A new_*() function, likely from the elmer package. Defaults to elmer::new_chat_claude(). To set a persistent alternative default, set the .clipal_fn option; see examples below.

...

Additional arguments to fn. The system_prompt argument will be ignored if supplied. To set persistent defaults, set the .clipal_args option; see examples below.

.ns

The package that the new_*() function is exported from.

Examples

if (FALSE) {
# to create a chat with claude:
clipal()

# or with OpenAI's 4o-mini:
clipal(
  "new_chat_openai",
  model = "gpt-4o-mini"
)

# to set OpenAI's 4o-mini as the default, for example, set the
# following options (possibly in your .Rprofile, if you'd like
# them to persist across sessions):
options(
  .clipal_fn = "new_chat_openai",
  .clipal_args = list(model = "gpt-4o-mini")
)
}